double >>>>

+++ atividades/Trabalho 04/Questão 01/Division.java
@@ -6,7 +6,8 @@ class Division implements Expression {
         this.right = r;
     }
 
-    public float calculate(){
+    
+    public double calculate(){
         return this.left.calculate() / this.right.calculate();
     }
 }
